refactor: entity 패키지 ui 패키지로부터 분리
