+++ app/src/main/java/woowacourse/movie/ui/fragment/movielist/HomeFragment.kt
@@ -9,9 +9,9 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.movie.R
+import woowacourse.movie.data.entity.Ads
+import woowacourse.movie.data.entity.Movies
 import woowacourse.movie.ui.activity.MovieDetailActivity
-import woowacourse.movie.ui.entity.Ads
-import woowacourse.movie.ui.entity.Movies
 import woowacourse.movie.ui.fragment.movielist.adapter.MovieAdapter
 import woowacourse.movie.ui.model.AdModel
 import woowacourse.movie.ui.model.MovieModel
