+++ app/src/main/java/woowacourse/movie/ui/fragment/reservationlist/ReservationListFragment.kt
@@ -9,8 +9,8 @@ import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.movie.R
+import woowacourse.movie.data.entity.Reservations
 import woowacourse.movie.ui.activity.MovieTicketActivity
-import woowacourse.movie.ui.entity.Reservations
 import woowacourse.movie.ui.fragment.reservationlist.adapter.ReservationAdapter
 import woowacourse.movie.ui.model.MovieTicketModel
